Restructured build system to support ti and imx  platforms
PROD-11253

+++ board/scripts-common/mksdcard.sh
@@ -71,8 +71,24 @@ BOOT_SIZE=48
 SWAP_SIZE=256
 PERM_SIZE=48
 
-{ [ -f "${SRCDIR}/rootfs.bin" ] || [ -f "${SRCDIR}/kernel.itb" ]; } && \
-	boot_only=false maxpart=6 || boot_only=true maxpart=1
+if [ ! -f "${SRCDIR}/rootfs.bin" ] && [ ! -f "${SRCDIR}/u-boot.itb" ]; then 
+	for f in "${SRCDIR}"/*.swu ; do
+		[ -f "${f}" ] || continue
+
+		TEMP_SRC=$(mktemp -d -t mksdcard.XXXXXX)
+		cpio -idmv < "${f}" -D "${TEMP_SRC}"
+		SRCDIR=${TEMP_SRC}
+		break
+	done
+fi
+
+if [ -f "${SRCDIR}/rootfs.bin" ]; then
+	boot_only=false 
+	maxpart=6
+else
+	boot_only=true
+	maxpart=1
+fi
 
 if ! ${boot_only}; then
 	# Calculate rootfs size
@@ -159,10 +175,18 @@ create_boot_partition() {
 	/usr/bin/mount "${1}" "${BOOT_PART}"
 
 	# Copy files to boot partition
-	cp -t "${BOOT_PART}" \
-		"${SRCDIR}/boot.${EXT}" \
-		"${SRCDIR}/u-boot.itb" \
-		"${SRCDIR}/uboot.env"
+	if [ -f "${SRCDIR}/tispl.bin" ]; then
+		cp -t "${BOOT_PART}" \
+			"${SRCDIR}/tispl.bin" \
+			"${SRCDIR}/tiboot3.bin" \
+			"${SRCDIR}/u-boot.img" \
+			"${SRCDIR}/uboot.env"
+	else
+		cp -t "${BOOT_PART}" \
+			"${SRCDIR}/boot.${EXT}" \
+			"${SRCDIR}/u-boot.itb" \
+			"${SRCDIR}/uboot.env"
+	fi
 
 	${boot_only} ||
 		cp -t "${BOOT_PART}" "${SRCDIR}/kernel.itb"
@@ -236,5 +260,7 @@ wait
 sync
 
 unmount_all "${TARGET}"
+
+[ ! -d "${TEMP_SRC}" ] || rm -rf "${TEMP_SRC}"
 
 echo "[Done]"
